fix10: 重复条目去重 + int溢出修复 + 文字溢出修复 + Footer截断修复
- 子仓列表按坐标去重(chunk卸载重建导致重复)
- eut全链路改long(修复Capacity显示-2147483648)
- machineMetaId改int(修复机器图标/名称错误)
- Flow加宽度约束(修复电网电量文字溢出)
- ListWidget高度188→170(修复Footer被背包遮挡)
- findAttachedMachine重写(多方块控制器优先)

+++ src/main/java/com/wztwzt/ae2_qof/hatch/adaptive/AdaptiveNetDynamoHatch.java
@@ -97,9 +97,9 @@ public class AdaptiveNetDynamoHatch extends MTEHatchDynamo {
             if (machineMTE != null) {
                 net.minecraft.item.ItemStack machineStack = machineMTE.getStackForm(1L);
                 if (machineStack != null) {
-                    helper.setMachineInfo((short) machineStack.getItemDamage(), machineStack.getDisplayName());
+                    helper.setMachineInfo(machineStack.getItemDamage(), machineStack.getDisplayName());
                 } else {
-                    helper.setMachineInfo((short) -1,
+                    helper.setMachineInfo(-1,
                         net.minecraft.util.StatCollector.translateToLocal(helper.getHatchType().getTranslationKey()));
                 }
             } else {
@@ -107,9 +107,9 @@ public class AdaptiveNetDynamoHatch extends MTEHatchDynamo {
                 if (mte != null) {
                     net.minecraft.item.ItemStack stack = mte.getStackForm(1L);
                     if (stack != null) {
-                        helper.setMachineInfo((short) stack.getItemDamage(), stack.getDisplayName());
+                        helper.setMachineInfo(stack.getItemDamage(), stack.getDisplayName());
                     } else {
-                        helper.setMachineInfo((short) -1,
+                        helper.setMachineInfo(-1,
                             net.minecraft.util.StatCollector.translateToLocal(helper.getHatchType().getTranslationKey()));
                     }
                 }
