+++ src/main/java/com/wztwzt/ae2_qof/hatch/adaptive/AdaptiveNetHatch.java
@@ -133,9 +133,9 @@ public class AdaptiveNetHatch extends MTEHatchEnergy {
             if (machineMTE != null) {
                 net.minecraft.item.ItemStack machineStack = machineMTE.getStackForm(1L);
                 if (machineStack != null) {
-                    helper.setMachineInfo((short) machineStack.getItemDamage(), machineStack.getDisplayName());
+                    helper.setMachineInfo(machineStack.getItemDamage(), machineStack.getDisplayName());
                 } else {
-                    helper.setMachineInfo((short) -1,
+                    helper.setMachineInfo(-1,
                         net.minecraft.util.StatCollector.translateToLocal(helper.getHatchType().getTranslationKey()));
                 }
             } else {
@@ -143,9 +143,9 @@ public class AdaptiveNetHatch extends MTEHatchEnergy {
                 if (mte != null) {
                     net.minecraft.item.ItemStack stack = mte.getStackForm(1L);
                     if (stack != null) {
-                        helper.setMachineInfo((short) stack.getItemDamage(), stack.getDisplayName());
+                        helper.setMachineInfo(stack.getItemDamage(), stack.getDisplayName());
                     } else {
-                        helper.setMachineInfo((short) -1,
+                        helper.setMachineInfo(-1,
                             net.minecraft.util.StatCollector.translateToLocal(helper.getHatchType().getTranslationKey()));
                     }
                 }
